app/requirements.txt
fastapi
python-multipart
pydantic<3.0
pydantic-settings
orjson
gunicorn
uvicorn
python-dotenv